services/messaging/requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.22.0
python-socketio==5.9.0
aiohttp==3.8.1
redis==4.5.0
python-dotenv==1.0.0